added DB connectivity

+++ src/main/java/com/springproject/model/CreateOrderResponse.kt
@@ -18,5 +18,5 @@ data class OrderPositionResponse(
         val id: String,
         val orderId: String,
         val productId:String,
-        val quantity: Long
+        val quantity: Int
 )
